Refactor: Move the ContentManager to the correct package
